+++ integration-tests/src/async_vault/constants.rs
@@ -1,3 +1,1 @@
-pub const VAULT_CONFIG_SEED: &[u8] = b"vault";
-pub const RESERVE_CONFIG_SEED: &[u8] = b"reserve";
-pub const PENDING_VAULT_SEED: &[u8] = b"pending";
+
