+++ integration-tests/src/async_vault/fee_extensions.rs
@@ -1,15 +1,19 @@
-use async_vault_client::{FeeType, Vault};
+use async_vault_client::{sdk::program_id, FeeType, Vault};
+use litesvm::LiteSVM;
 use solana_sdk::{account::ReadableAccount, signature::Keypair, signer::Signer};
 
-use crate::async_vault::helper_functions::{
-    assert_error_code, init_deposit_fee, init_withdrawal_fee, new_svm, setup_vault,
+use crate::helper_functions::{
+    assert_error_code, init_deposit_fee, init_withdrawal_fee, setup_async_vault,
     update_deposit_fee, update_withdrawal_fee,
 };
 
 #[test]
 fn test_initialize_and_update_deposit_fee() {
-    let mut svm = new_svm();
-    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_vault(&mut svm);
+    let mut svm = LiteSVM::new();
+
+    let program_bytes = include_bytes!("../../../target/deploy/async_vault.so");
+    svm.add_program(program_id(), program_bytes).unwrap();
+    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_async_vault(&mut svm);
 
     let deposit_fee = FeeType::FixedAmount { amount: 100 };
     init_deposit_fee(
@@ -40,8 +44,11 @@ fn test_initialize_and_update_deposit_fee() {
 
 #[test]
 fn test_initialize_and_update_withdrawal_fee() {
-    let mut svm = new_svm();
-    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_vault(&mut svm);
+    let mut svm = LiteSVM::new();
+
+    let program_bytes = include_bytes!("../../../target/deploy/async_vault.so");
+    svm.add_program(program_id(), program_bytes).unwrap();
+    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_async_vault(&mut svm);
 
     let withdrawal_fee = FeeType::Percentage { bps: 200 };
     init_withdrawal_fee(
@@ -70,8 +77,11 @@ fn test_initialize_and_update_withdrawal_fee() {
 
 #[test]
 fn test_initialize_both_fees() {
-    let mut svm = new_svm();
-    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_vault(&mut svm);
+    let mut svm = LiteSVM::new();
+
+    let program_bytes = include_bytes!("../../../target/deploy/async_vault.so");
+    svm.add_program(program_id(), program_bytes).unwrap();
+    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_async_vault(&mut svm);
 
     let deposit_fee = FeeType::FixedAmount { amount: 100 };
     init_deposit_fee(
@@ -100,8 +110,11 @@ fn test_initialize_both_fees() {
 
 #[test]
 fn test_initialize_deposit_fee_duplicate_fails() {
-    let mut svm = new_svm();
-    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_vault(&mut svm);
+    let mut svm = LiteSVM::new();
+
+    let program_bytes = include_bytes!("../../../target/deploy/async_vault.so");
+    svm.add_program(program_id(), program_bytes).unwrap();
+    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_async_vault(&mut svm);
 
     let deposit_fee = FeeType::FixedAmount { amount: 100 };
     init_deposit_fee(
@@ -127,8 +140,11 @@ fn test_initialize_deposit_fee_duplicate_fails() {
 
 #[test]
 fn test_initialize_withdrawal_fee_duplicate_fails() {
-    let mut svm = new_svm();
-    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_vault(&mut svm);
+    let mut svm = LiteSVM::new();
+
+    let program_bytes = include_bytes!("../../../target/deploy/async_vault.so");
+    svm.add_program(program_id(), program_bytes).unwrap();
+    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_async_vault(&mut svm);
 
     let withdrawal_fee = FeeType::Percentage { bps: 100 };
     init_withdrawal_fee(
@@ -154,8 +170,11 @@ fn test_initialize_withdrawal_fee_duplicate_fails() {
 
 #[test]
 fn test_update_deposit_fee_before_init_fails() {
-    let mut svm = new_svm();
-    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_vault(&mut svm);
+    let mut svm = LiteSVM::new();
+
+    let program_bytes = include_bytes!("../../../target/deploy/async_vault.so");
+    svm.add_program(program_id(), program_bytes).unwrap();
+    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_async_vault(&mut svm);
 
     let deposit_fee = FeeType::FixedAmount { amount: 100 };
     let result = update_deposit_fee(
@@ -170,8 +189,11 @@ fn test_update_deposit_fee_before_init_fails() {
 
 #[test]
 fn test_update_withdrawal_fee_before_init_fails() {
-    let mut svm = new_svm();
-    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_vault(&mut svm);
+    let mut svm = LiteSVM::new();
+
+    let program_bytes = include_bytes!("../../../target/deploy/async_vault.so");
+    svm.add_program(program_id(), program_bytes).unwrap();
+    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_async_vault(&mut svm);
 
     let withdrawal_fee = FeeType::Percentage { bps: 100 };
     let result = update_withdrawal_fee(
@@ -186,8 +208,11 @@ fn test_update_withdrawal_fee_before_init_fails() {
 
 #[test]
 fn test_initialize_deposit_fee_invalid_bps_fails() {
-    let mut svm = new_svm();
-    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_vault(&mut svm);
+    let mut svm = LiteSVM::new();
+
+    let program_bytes = include_bytes!("../../../target/deploy/async_vault.so");
+    svm.add_program(program_id(), program_bytes).unwrap();
+    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_async_vault(&mut svm);
 
     let deposit_fee = FeeType::Percentage { bps: 10_001 };
     let result = init_deposit_fee(
@@ -202,8 +227,11 @@ fn test_initialize_deposit_fee_invalid_bps_fails() {
 
 #[test]
 fn test_initialize_withdrawal_fee_invalid_bps_fails() {
-    let mut svm = new_svm();
-    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_vault(&mut svm);
+    let mut svm = LiteSVM::new();
+
+    let program_bytes = include_bytes!("../../../target/deploy/async_vault.so");
+    svm.add_program(program_id(), program_bytes).unwrap();
+    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_async_vault(&mut svm);
 
     let withdrawal_fee = FeeType::Percentage { bps: 10_001 };
     let result = init_withdrawal_fee(
@@ -218,8 +246,11 @@ fn test_initialize_withdrawal_fee_invalid_bps_fails() {
 
 #[test]
 fn test_initialize_fee_unauthorized_signer_fails() {
-    let mut svm = new_svm();
-    let (_, _, _, share_mint, _, _, vault_pubkey) = setup_vault(&mut svm);
+    let mut svm = LiteSVM::new();
+
+    let program_bytes = include_bytes!("../../../target/deploy/async_vault.so");
+    svm.add_program(program_id(), program_bytes).unwrap();
+    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_async_vault(&mut svm);
 
     let unauthorized = Keypair::new();
     svm.airdrop(&unauthorized.pubkey(), 1_000_000_000).unwrap();
@@ -237,8 +268,11 @@ fn test_initialize_fee_unauthorized_signer_fails() {
 
 #[test]
 fn test_update_fee_unauthorized_signer_fails() {
-    let mut svm = new_svm();
-    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_vault(&mut svm);
+    let mut svm = LiteSVM::new();
+
+    let program_bytes = include_bytes!("../../../target/deploy/async_vault.so");
+    svm.add_program(program_id(), program_bytes).unwrap();
+    let (authority, _, _, share_mint, _, _, vault_pubkey) = setup_async_vault(&mut svm);
 
     let deposit_fee = FeeType::FixedAmount { amount: 100 };
     init_deposit_fee(
