+++ integration-tests/src/async_vault/mod.rs
@@ -1,4 +1,3 @@
 mod constants;
 mod create_vault;
 mod fee_extensions;
-mod helper_functions;
