+++ integration-tests/src/lib.rs
@@ -1,4 +1,6 @@
 #[cfg(test)]
 mod async_vault;
 #[cfg(test)]
+pub(crate) mod helper_functions;
+#[cfg(test)]
 mod vault;
