+++ integration-tests/src/vault/mint.rs
@@ -7,7 +7,7 @@ use solana_sdk::{account::ReadableAccount, signature::Keypair, signer::Signer};
 use test_case::test_case;
 use vault_client::{sdk::program_id, FeeType, Pubkey, Vault};
 
-use crate::vault::helper_functions::{
+use crate::helper_functions::{
     assert_error_code, create_ata, create_mint, create_mint_with_transfer_fee, get_fee,
     get_mint_supply, get_token_account_amount, get_vault_asset_balance, helper_mint_to, mint,
     recv_amount_from_params, set_up_vault,
