+++ integration-tests/src/vault/mod.rs
@@ -4,7 +4,6 @@ mod create_vault;
 mod deposit;
 mod deposit_hook;
 mod fee_extensions;
-mod helper_functions;
 mod mint;
 mod redeem;
 mod update_vault;
