+++ programs/vault/src/error.rs
@@ -79,4 +79,7 @@ pub enum VaultProgramError {
 
     #[msg("Mints should be different.")]
     MintsShouldBeDifferent,
+
+    #[msg("Share mint supply should be zero.")]
+    ShareMintSupplyShouldBeZero,
 }
